// shared/src/types.ts
export interface IEducation {
  _id: string;
  institution: string;
  degree: string;
  field: string;
  location: string;
  startDate: Date;
  endDate?: Date;
  current: boolean;
  grade?: string;
  description?: string;
  order: number;
  visibility: boolean;
  createdAt: Date;
  updatedAt: Date;
  createdBy: string;
}